requirements.txt
pyTelegramBotAPI==4.15.1
python-dotenv
Pillow
Flask
gunicorn
